refac(circom): move `VerifyingKey` to `zkey/verifying_key.h`
